fix: keep resume path repair focused
