+++ cmuxTests/SessionPersistenceResumeBindingTests.swift
@@ -1,0 +1,67 @@
+import XCTest
+
+#if canImport(cmux_DEV)
+@testable import cmux_DEV
+#elseif canImport(cmux)
+@testable import cmux
+#endif
+
+extension SessionPersistenceTests {
+    func testAgentHookSurfaceResumeStartupInputFallsBackWhenRecordedAgentExecutableMoved() throws {
+        let fileManager = FileManager.default
+        let root = fileManager.temporaryDirectory
+            .appendingPathComponent("cmux-surface-resume-moved-agent-\(UUID().uuidString)", isDirectory: true)
+        let bin = root.appendingPathComponent("bin", isDirectory: true)
+        let cwd = root.appendingPathComponent("repo", isDirectory: true)
+        let missingExecutable = root
+            .appendingPathComponent("deleted-node", isDirectory: true)
+            .appendingPathComponent("bin", isDirectory: true)
+            .appendingPathComponent("codex", isDirectory: false)
+        let outputURL = root.appendingPathComponent("codex-output.txt", isDirectory: false)
+        try fileManager.createDirectory(at: bin, withIntermediateDirectories: true)
+        try fileManager.createDirectory(at: cwd, withIntermediateDirectories: true)
+        defer { try? fileManager.removeItem(at: root) }
+
+        let fakeCodex = bin.appendingPathComponent("codex", isDirectory: false)
+        try """
+        #!/bin/sh
+        printf '%s|%s\\n' "$PWD" "$*" > "$CMUX_FAKE_CODEX_OUTPUT"
+        """.write(to: fakeCodex, atomically: true, encoding: .utf8)
+        try fileManager.setAttributes([.posixPermissions: 0o700], ofItemAtPath: fakeCodex.path)
+
+        let quotedCwd = "'\(cwd.path)'"
+        let binding = SurfaceResumeBindingSnapshot(
+            kind: "codex",
+            command: "{ cd -- \(quotedCwd) 2>/dev/null || [ ! -d \(quotedCwd) ]; } && "
+                + "'\(missingExecutable.path)' 'resume' 'session-moved-cli' '--yolo'",
+            cwd: cwd.path,
+            checkpointId: "session-moved-cli",
+            source: "agent-hook",
+            autoResume: true
+        )
+        let startupInput = try XCTUnwrap(binding.startupInput)
+
+        let process = Process()
+        process.executableURL = URL(fileURLWithPath: "/bin/zsh")
+        process.arguments = ["-fc", startupInput]
+        process.environment = [
+            "PATH": "\(bin.path):/usr/bin:/bin",
+            "CMUX_FAKE_CODEX_OUTPUT": outputURL.path,
+        ]
+        let stderr = Pipe()
+        process.standardError = stderr
+
+        try process.run()
+        process.waitUntilExit()
+
+        let errorText = String(
+            data: stderr.fileHandleForReading.readDataToEndOfFile(),
+            encoding: .utf8
+        ) ?? ""
+        XCTAssertEqual(process.terminationStatus, 0, errorText)
+
+        let output = try String(contentsOf: outputURL, encoding: .utf8)
+        XCTAssertEqual(output, "\(cwd.path)|resume session-moved-cli --yolo\n")
+        XCTAssertFalse(startupInput.contains(missingExecutable.path), startupInput)
+    }
+}
